// server/models/sentOasysRiskInformation.ts
export interface SentOasysRiskInformation {
  riskSummaryWhoIsAtRisk: string | null
  riskSummaryNatureOfRisk: string | null
  riskSummaryRiskImminence: string | null
  riskToSelfSuicide: string | null
  riskToSelfSelfHarm: string | null
  riskToSelfHostelSetting: string | null
  riskToSelfVulnerability: string | null
  additionalInformation: string | null
}